Move environment variables into layer creator
This commit moves where the environment variables are being set to inside of the layer creator. The environment variables are attached to the layer, so they should be set in the layer creator. That way they are all of the actions happen when the layer is recrated and only when the layer is recreated. The environment variables do not need to be set every time the contributor runs, just when the layer metadata changes.

Also, cleans up some ioutil calls.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>