Remove process type
This PR removes the process type contributed by the health checker buildpack. This was problematic in that it would trigger the exec.d helpers to run every time the health check was run, which caused issues.

Instead, we install the health check binary, print the path at which it's installed and one can invoke it directly. Then we attempt to create a symlink to the health check binary under `/workspace/health-check`. This is then the shortcut that one can use to easily invoke the health check binary.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>