Remove setting env variables
It is not possible to set env variables because we are no longer using the launcher to start the process. We can't use the launcher because it will run the exec.d helpers, so that means no env variables baked into the image. Users must set them now at runtime.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>